Endpoint size globals in the host demonstration applications are now 16-bit, as they should be.
Removed unneeded Endpoint_Ignore_Word command from the control command handler event routine in the Mass Storage demo.

If the logical block address is out of range, the correct LOGICAL BLOCK OUT OF RANGE sense key is returned instead of the incorrect HARDWARE ERROR key.

Inquiry command now uses 16-bits for the allocation size, to be able to hold the maximum allocation size given by the host.

git-svn-id: http://lufa-lib.googlecode.com/svn@123 d5102386-fcda-11dd-9fdb-3debd5008f28

+++ Demos/KeyboardHost/KeyboardHost.c
@@ -41,8 +41,8 @@ TASK_LIST
 };
 
 /* Globals */
-uint8_t KeyboardDataEndpointNumber;
-uint8_t KeyboardDataEndpointSize;
+uint8_t  KeyboardDataEndpointNumber;
+uint16_t KeyboardDataEndpointSize;
 
 int main(void)
 {
