+++ Demos/MouseHost/MouseHost.c
@@ -42,8 +42,8 @@ TASK_LIST
 };
 
 /* Globals */
-uint8_t MouseDataEndpointNumber;
-uint8_t MouseDataEndpointSize;
+uint8_t  MouseDataEndpointNumber;
+uint16_t MouseDataEndpointSize;
 
 int main(void)
 {
